requirements.txt
streamlit
google-generativeai # Generative AI
python-dotenv # Meant for local environme

